feedback ao usuário

+++ src/app/pages/pesquisadores/consulta-pesquisadores/consulta-pesquisadores.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { PesquisadoresService } from '../pesquisadores.service';
 import { Pesquisador } from 'src/app/models/pesquisador';
+import { UtilsService } from 'src/app/utils/utils.service';
 
 @Component({
   selector: 'consulta-pesquisadores',
@@ -9,16 +10,19 @@ import { Pesquisador } from 'src/app/models/pesquisador';
 export class ConsultaPesquisadoresComponent implements OnInit {
 
   pesquisadores: Pesquisador[];
+  isLoading: boolean;
 
-  constructor(private pesquisadoresService: PesquisadoresService) { }
+  constructor(private pesquisadoresService: PesquisadoresService, private utils: UtilsService) { }
 
   ngOnInit() {
     this.getPesquisadores();
   }
 
   getPesquisadores() {
+    this.isLoading = true;
     this.pesquisadoresService.getPesquisadores()
       .subscribe((res: Pesquisador[]) => {
+        this.isLoading = false;
         this.pesquisadores = res;
       });
   }
@@ -32,8 +36,13 @@ export class ConsultaPesquisadoresComponent implements OnInit {
   }
 
   excluirPesquisador(id) {
-    this.pesquisadoresService.excluirPesquisador(id).subscribe(() => {
-      console.log('excluido');
+    this.utils.confirm('Confirmar', 'Você realmente deseja excluir este registro? Essa ação é irreversível', () => {
+      this.isLoading = true;
+      this.pesquisadoresService.excluirPesquisador(id).subscribe(() => {
+        this.isLoading = false;
+        this.utils.success('Sucesso!', 'O pesquisador foi excluído com sucesso.', this.getPesquisadores());
+        console.log('excluido');
+      });
     });
   }
